Refactor funcinfo to funcInfo as intended by old readme
