+++ src/func-info.h
